collocate tests under __tests__
