+++ src/__tests__/App.test.js
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { render } from 'enzyme';
 import configureMockStore from 'redux-mock-store';
-import App from '../src/App';
+import App from 'App';
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
