+++ src/components/Board/__tests__/fixtures/initialPlacementOnly.json
