+++ src/components/Board/__tests__/index.test.js
@@ -9,14 +9,6 @@ import rootReducer from 'reducers';
 import Board from 'components/Board';
 
 import initialPlacementOnly from './fixtures/initialPlacementOnly.json';
-import {
-  clickSquare,
-  expectToBeEmptySquare,
-  expectToHavePiece,
-  expectSquaresToBeInLastMove,
-  expectSquaresToBeSelected,
-  expectSquaresToBeTargeted,
-} from './helpers';
 
 const initialState = {
   gameSlug: null,
@@ -46,6 +38,48 @@ const getBoard = store => (
     <Board />
   </Provider>
 );
+
+// Helper functions
+
+const getSquareNodes = (wrapper, squares) =>
+  wrapper.find('Square').findWhere(node => squares.includes(node.key()));
+
+const getSquareNode = (wrapper, square) => getSquareNodes(wrapper, [square]);
+
+const clickSquare = (wrapper, square) => {
+  getSquareNode(wrapper, square)
+    .children()
+    .simulate('click');
+
+  jest.runOnlyPendingTimers();
+
+  wrapper.update();
+};
+
+const expectSquaresToBeSelected = (wrapper, squares) => {
+  const nodes = getSquareNodes(wrapper, squares).find('SelectionIndicator');
+  expect(nodes).toHaveLength(squares.length);
+};
+
+const expectSquaresToBeTargeted = (wrapper, squares) => {
+  const nodes = getSquareNodes(wrapper, squares).find('TargetIndicator');
+  expect(nodes).toHaveLength(squares.length);
+};
+
+const expectSquaresToBeInLastMove = (wrapper, squares) => {
+  const nodes = getSquareNodes(wrapper, squares).find('LastMoveIndicator');
+  expect(nodes).toHaveLength(squares.length);
+};
+
+const expectToHavePiece = (wrapper, square, piece) => {
+  const node = getSquareNode(wrapper, square).find(`.Piece .${piece}`);
+  expect(node).toHaveLength(1);
+};
+
+const expectToBeEmptySquare = (wrapper, square) => {
+  const node = getSquareNode(wrapper, square).find('.Piece');
+  expect(node).toHaveLength(0);
+};
 
 describe('Board', () => {
   test('renders without crashing', () => {
