+++ src/reducers/__tests__/selectedSquare.test.js
