+++ src/services/logic/__tests__/square.test.js
